don't finish

+++ src/components/layout/header-home.tsx
@@ -4,11 +4,13 @@ import { useEffect, useState } from "react";
 import { Toaster, toast } from "sonner";
 import { useRouter } from "next/navigation";
 import { ShoppingCartOutlined } from "@ant-design/icons";
-import { Button } from "antd";
+import { Button, Input } from "antd";
 import { CartModel } from "@/models/cartmodel";
-import CartDrawer from "./cartDrawer";
+import CartDrawer from "@/components/ui/CartDrawer";
+import Search from "antd/es/transfer/search";
+import { SearchProps } from "antd/es/input";
 
-export default function NavHome() {
+export default function HeaderHome() {
     const [userName, setUserName] = useState("");
     const [email, setEmail] = useState("");
     const [userId, setUserId] = useState("");
@@ -16,7 +18,10 @@ export default function NavHome() {
     const [datacart, setDatacart] = useState<CartModel[]>([]);
     const [open, setOpen] = useState<boolean>(false);
     const [loading, setLoading] = useState<boolean>(true);
-
+    const isLogin = useState<boolean>(false);
+    const { Search } = Input;
+    const onSearch: SearchProps["onSearch"] = (value, _e, info) =>
+        console.log(info?.source, value);
     useEffect(() => {
         const cartData = localStorage.getItem("CartData");
         const storedUserId = localStorage.getItem("userId");
@@ -28,6 +33,7 @@ export default function NavHome() {
         }
         if (storedUserId) {
             setUserId(storedUserId);
+            isLogin[0] = true;
         }
 
         if (storedEmail) {
@@ -42,7 +48,7 @@ export default function NavHome() {
     const handerLogout = async () => {
         try {
             const response = await fetch(
-                `https://localhost:44372/api/Account/logout?userEmail=${email}`,
+                `${process.env.API_URL}Account/logout?userEmail=${email}`,
                 {
                     method: "GET",
                     headers: {
@@ -56,19 +62,12 @@ export default function NavHome() {
                 router.push("/");
                 router.refresh();
                 localStorage.clear();
-                // localStorage.removeItem("access_token");
-                // localStorage.removeItem("email");
-                // localStorage.removeItem("name");
-                // localStorage.removeItem("CartData");
-                // localStorage.removeItem("userId");
             }
         } catch (error) {
             alert("Logout failed");
             console.error("Logout error:", error);
         }
     };
-    // console.log(email);
-
     const showLoading = () => {
         setOpen(true);
         setLoading(true);
@@ -86,27 +85,25 @@ export default function NavHome() {
 
     const deleteCartItem = async (userId?: string, itemId?: string) => {
         try {
-            const response = await fetch(
-                `https://localhost:44372/api/Cart/${localStorage.getItem(
-                    "userId"
-                )}/item/${itemId}`,
-                {
-                    method: "DELETE",
-                    headers: {
-                        "Content-Type": "application/json",
-                        Authorization: `Bearer ${
-                            localStorage.getItem("access_token") ?? ""
-                        }`,
-                    },
+            if (isLogin[0]) {
+                const response = await fetch(
+                    `${process.env.API_URL}Cart/${userId}/item/${itemId}`,
+                    {
+                        method: "DELETE",
+                        headers: {
+                            "Content-Type": "application/json",
+                            Authorization: `Bearer ${
+                                localStorage.getItem("access_token") ?? ""
+                            }`,
+                        },
+                    }
+                );
+
+                if (!response.ok) {
+                    throw new Error("Failed to delete item");
                 }
-            );
-
-            if (!response.ok) {
-                throw new Error("Failed to delete item");
+                return response.json();
             }
-            console.log("SUCCESS");
-
-            return response.json();
         } catch (error) {
             console.error("Error deleting item:", error);
             throw error;
@@ -125,12 +122,37 @@ export default function NavHome() {
     };
 
     return (
-        <div>
-            <div className="flex py-5 bg-white shadow-xl border-y-2">
-                <Toaster position="top-right" richColors duration={2000} />
+        <nav className="flex py-5 bg-white shadow-xl border-y-2">
+            <Toaster position="top-right" richColors duration={2000} />
+            <div className="max-w-screen-2xl flex flex-wrap items-center justify-between mx-auto p-3 w-full">
                 <div className="basis-1/4 text-center font-extralight text-zinc-700">
                     <Link href="/">TV FURNITURE</Link>{" "}
                 </div>
+                <button
+                    data-collapse-toggle="navbar-default"
+                    type="button"
+                    className="inline-flex items-center p-2 w-10 h-10 justify-center text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600"
+                    aria-controls="navbar-default"
+                    aria-expanded="false"
+                >
+                    <span className="sr-only">Open main menu</span>
+                    <svg
+                        className="w-5 h-5"
+                        aria-hidden="true"
+                        xmlns="http://www.w3.org/2000/svg"
+                        fill="none"
+                        viewBox="0 0 17 14"
+                    >
+                        <path
+                            stroke="currentColor"
+                            strokeLinecap="round"
+                            strokeLinejoin="round"
+                            strokeWidth="2"
+                            d="M1 1h15M1 7h15M1 13h15"
+                        />
+                    </svg>
+                </button>
+
                 <div className="basis-1/2 flex">
                     <ul className=" flex">
                         <li>
@@ -166,10 +188,11 @@ export default function NavHome() {
                             </Link>
                         </li>
                         <li>
-                            <input
-                                placeholder="Tìm sản phẩm"
-                                type="text"
-                                className="ml-2 border rounded-xl text-sm px-2 py-0.5 bg-slate-100"
+                            <Search
+                                placeholder="Tìm kiếm sản phẩm"
+                                allowClear
+                                onSearch={onSearch}
+                                style={{ width: 220 }}
                             />
                         </li>
                     </ul>
@@ -225,74 +248,10 @@ export default function NavHome() {
                                     handleContinueBuyProduct
                                 }
                             ></CartDrawer>
-                            {/* <Drawer
-                                closable
-                                destroyOnClose
-                                title={<p>Giỏ hàng</p>}
-                                placement="right"
-                                open={open}
-                                loading={loading}
-                                onClose={() => setOpen(false)}
-                            >
-                                {datacart?.map((item) => (
-                                    <div className="w-full flex">
-                                        <div className="w-1/4">
-                                            <img
-                                                src={item?.images?.[0]?.url}
-                                                alt=""
-                                            />
-                                        </div>
-                                        <div className="w-3/4 flex justify-between items-center">
-                                            <div>
-                                                <div>{item?.name}</div>
-                                                <div>
-                                                    {item?.price} <span>x</span>{" "}
-                                                    {item?.quantity}
-                                                </div>
-                                            </div>
-                                            <button
-                                                onClick={() =>
-                                                    handleRemoveItem(
-                                                        userId,
-                                                        item.id
-                                                    )
-                                                }
-                                            >
-                                                <FontAwesomeIcon
-                                                    icon={faTimes}
-                                                />
-                                            </button>
-                                        </div>
-                                    </div>
-                                ))}
-                                <div className="flex justify-between items-center mt-4">
-                                    <div className="text-lg font-bold">
-                                        Tổng cộng: 11đ
-                                    </div>
-                                </div>
-                                <div className="mt-4">
-                                    <Button
-                                        type="primary"
-                                        style={{
-                                            marginBottom: 16,
-                                            marginRight: 34,
-                                        }}
-                                        onClick={handleContinueBuyProduct}
-                                    >
-                                        Xem giỏ hàng
-                                    </Button>
-                                    <Button
-                                        type="primary"
-                                        // onClick={handleCheckout}
-                                    >
-                                        Thanh toán
-                                    </Button>
-                                </div>
-                            </Drawer> */}
                         </li>
                     </ul>
                 </div>
             </div>
-        </div>
+        </nav>
     );
 }
